We need to change IP hader to fool loopback through our devices
IP address we need to change so that it comes back through our driver
else will loopback.

Also we need to disable ARP if we want packet to come to our device
driver in first place. Then also we need to provide header_ops so that
ETH header can be built properly with destination address for our mac
devices filled correctly within the driver.

Signed-off-by: Shobhit Kumar <kumar@shobhit.info>